backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
requests
faiss-cpu
pypdf
PyMuPDF
chromadb
python-multipart
sqlalchemy
aiosqlite
python-dotenv
python-docx
openpyxl
python-pptx
